requirements.txt
git+https://github.com/mmabrouk/chatgpt-wrapper
python-dotenv